server/requirements.txt
# Server dependencies
flask>=3.0.0
flask-cors>=4.0.0
flask-sqlalchemy>=3.1.0
Flask-Migrate>=4.0.0
cryptography>=41.0.0
python-dotenv>=1.0.0
werkzeug>=3.0.0
gunicorn>=21.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
pyjwt>=2.8.0
psutil>=5.9.0
flask-talisman>=1.0.0

# Optional: WebSocket support (uncomment if needed)
# flask-socketio>=5.3.0

# Optional: Video processing (uncomment if needed)
# opencv-python>=4.8.0
# numpy>=1.24.0
